Move login actions to a properly-named file
