+++ src/components/scenes/LoginScene.js
@@ -9,10 +9,10 @@ import { checkVersion } from 'react-native-check-version'
 import { getBundleId } from 'react-native-device-info'
 
 import ENV from '../../../env.json'
+import { initializeAccount, logoutRequest } from '../../actions/LoginActions.js'
 import edgeBackgroundImage from '../../assets/images/edgeBackground/login_bg.gif'
 import edgeLogo from '../../assets/images/edgeLogo/Edge_logo_L.png'
 import s from '../../locales/strings.js'
-import { initializeAccount, logoutRequest } from '../../modules/Login/action.js'
 import THEME from '../../theme/variables/airbitz.js'
 import { type DeepLink } from '../../types/DeepLink.js'
 import { connect } from '../../types/reactRedux.js'
