+++ src/modules/UI/components/ControlPanel/Component/Main.js
@@ -6,6 +6,7 @@ import { Actions } from 'react-native-router-flux'
 import Share from 'react-native-share'
 import { sprintf } from 'sprintf-js'
 
+import { logoutRequest } from '../../../../../actions/LoginActions.js'
 import { selectWalletFromModal } from '../../../../../actions/WalletActions.js'
 import buysellIcon from '../../../../../assets/images/sidenav/buysell.png'
 import exchangeIcon from '../../../../../assets/images/sidenav/exchange.png'
@@ -40,7 +41,6 @@ import s from '../../../../../locales/strings.js'
 import { THEME } from '../../../../../theme/variables/airbitz.js'
 import { connect } from '../../../../../types/reactRedux.js'
 import { scale } from '../../../../../util/scaling.js'
-import { logoutRequest } from '../../../../Login/action.js'
 import styles from '../style'
 import { Button } from './Button/Button.ui.js'
 import { Separator } from './Separator/Separator.ui.js'
