+++ src/modules/UI/components/ControlPanel/Component/UserList.js
@@ -7,10 +7,10 @@ import { Alert, ScrollView, TouchableHighlight, View } from 'react-native'
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons'
 import { sprintf } from 'sprintf-js'
 
+import { logoutRequest } from '../../../../../actions/LoginActions.js'
 import { showError } from '../../../../../components/services/AirshipInstance.js'
 import s from '../../../../../locales/strings'
 import { connect } from '../../../../../types/reactRedux.js'
-import { logoutRequest } from '../../../../Login/action.js'
 import T from '../../../components/FormattedText/FormattedText.ui.js'
 import { deleteLocalAccount } from '../action.js'
 import styles from '../style'
